pyproject.toml
[project]
name = "chat-bots-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot==13.15",
    "requests>=2.32.3",
    "ruff>=0.11.9",
    "standard-imghdr>=3.13.0",
    "urllib3==1.26.15",
]
